Extracted a parent SensorService
