+++ app/src/main/java/io/lunarlogic/aircasting/sensor/SensorService.kt
@@ -1,0 +1,55 @@
+package io.lunarlogic.aircasting.sensor
+
+import android.R
+import android.app.*
+import android.content.Context
+import android.content.Intent
+import android.os.Build
+import android.os.IBinder
+import androidx.core.app.NotificationCompat
+import androidx.core.content.ContextCompat
+import io.lunarlogic.aircasting.screens.main.MainActivity
+
+
+abstract class SensorService : Service() {
+    private val CHANNEL_ID = "Aircasting ForegroundService"
+
+    override fun onBind(intent: Intent?): IBinder? {
+        return null
+    }
+
+    override fun onStartCommand(intent: Intent?, flags: Int, startId: Int): Int {
+        println("MARYSIA: microphpone start? onStartCommand")
+       startSensor()
+        val input = intent?.getStringExtra("inputExtra")
+        createNotificationChannel()
+        val notificationIntent = Intent(this, MainActivity::class.java)
+        val pendingIntent = PendingIntent.getActivity(
+            this,
+            0, notificationIntent, 0
+        )
+        val notification = NotificationCompat.Builder(this, CHANNEL_ID)
+            .setContentTitle("Aircasting: Microphone Service")
+            .setContentText(input)
+            .setSmallIcon(R.drawable.ic_btn_speak_now)
+            .setContentIntent(pendingIntent)
+            .build()
+
+        startForeground(1, notification)
+
+        return START_NOT_STICKY
+    }
+
+    abstract fun startSensor()
+
+    private fun createNotificationChannel() {
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+            val serviceChannel = NotificationChannel(
+                CHANNEL_ID, "Foreground Service Channel",
+                NotificationManager.IMPORTANCE_DEFAULT
+            )
+            val manager = getSystemService(NotificationManager::class.java)
+            manager!!.createNotificationChannel(serviceChannel)
+        }
+    }
+}
